requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-huggingface
langchain-community
pypdf
faiss-cpu
python-dotenv